Extract @voyantjs/templating; purify legal-contracts off @voyantjs/utils
Move the Liquid/Mustache template renderer + syntax validator out of
@voyantjs/utils into a new lean @voyantjs/templating package (liquidjs only).
@voyantjs/utils/template-renderer re-exports it, so finance/products/legal
runtime import paths are unchanged.

legal-contracts validates contract-body Liquid syntax via @voyantjs/templating
now, dropping its transitive dependency on @voyantjs/utils (and thus drizzle,
pdf-lib, and @voyantjs/db). Its tree is now zod + schema-kit + templating only.
