+++ packages/templating/package.json
@@ -1,12 +1,11 @@
 {
-  "name": "@voyantjs/legal-contracts",
+  "name": "@voyantjs/templating",
   "version": "0.96.0",
   "license": "Apache-2.0",
   "type": "module",
   "exports": {
     ".": "./src/index.ts",
-    "./contracts/validation": "./src/contracts/validation.ts",
-    "./policies/validation": "./src/policies/validation.ts"
+    "./template-renderer": "./src/template-renderer.ts"
   },
   "scripts": {
     "typecheck": "tsc --noEmit",
@@ -27,24 +26,17 @@
         "import": "./dist/index.js",
         "default": "./dist/index.js"
       },
-      "./contracts/validation": {
-        "types": "./dist/contracts/validation.d.ts",
-        "import": "./dist/contracts/validation.js",
-        "default": "./dist/contracts/validation.js"
-      },
-      "./policies/validation": {
-        "types": "./dist/policies/validation.d.ts",
-        "import": "./dist/policies/validation.js",
-        "default": "./dist/policies/validation.js"
+      "./template-renderer": {
+        "types": "./dist/template-renderer.d.ts",
+        "import": "./dist/template-renderer.js",
+        "default": "./dist/template-renderer.js"
       }
     },
     "main": "./dist/index.js",
     "types": "./dist/index.d.ts"
   },
   "dependencies": {
-    "@voyantjs/schema-kit": "workspace:*",
-    "@voyantjs/templating": "workspace:*",
-    "zod": "^4.3.6"
+    "liquidjs": "^10.24.0"
   },
   "devDependencies": {
     "@voyantjs/voyant-typescript-config": "workspace:*",
@@ -54,6 +46,6 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/voyantjs/voyant.git",
-    "directory": "packages/legal-contracts"
+    "directory": "packages/templating"
   }
 }
