+++ packages/utils/package.json
@@ -124,7 +124,8 @@
     "drizzle-orm": "^0.45.2",
     "liquidjs": "^10.24.0",
     "pdf-lib": "^1.17.1",
-    "zod": "^4.3.6"
+    "zod": "^4.3.6",
+    "@voyantjs/templating": "workspace:*"
   },
   "devDependencies": {
     "@voyantjs/voyant-typescript-config": "workspace:*",
